encapsulate ActivationStatus
