+++ web/src/SourcegraphWebApp.tsx
@@ -1,18 +1,16 @@
 import { ShortcutProvider } from '@slimsag/react-shortcuts'
-import H from 'history'
 import ServerIcon from 'mdi-react/ServerIcon'
 import * as React from 'react'
 import { Route } from 'react-router'
 import { BrowserRouter } from 'react-router-dom'
-import { combineLatest, from, fromEventPattern, Observable, Subscription } from 'rxjs'
-import { first, map, startWith } from 'rxjs/operators'
-import { ActivationCompleted, ActivationStep } from '../../shared/src/components/activation/Activation'
+import { combineLatest, from, fromEventPattern, of, Subscription } from 'rxjs'
+import { map, startWith, switchMap } from 'rxjs/operators'
+import { Activation, ActivationProps } from '../../shared/src/components/activation/Activation'
 import { setLinkComponent } from '../../shared/src/components/Link'
 import {
     createController as createExtensionsController,
     ExtensionsControllerProps,
 } from '../../shared/src/extensions/controller'
-import { dataOrThrowErrors, gql } from '../../shared/src/graphql/graphql'
 import * as GQL from '../../shared/src/graphql/schema'
 import { Notifications } from '../../shared/src/notifications/Notifications'
 import { PlatformContextProps } from '../../shared/src/platform/context'
@@ -20,7 +18,6 @@ import { EMPTY_SETTINGS_CASCADE, SettingsCascadeProps } from '../../shared/src/s
 import { TelemetryContext } from '../../shared/src/telemetry/telemetryContext'
 import { isErrorLike } from '../../shared/src/util/errors'
 import { authenticatedUser } from './auth'
-import { queryGraphQL } from './backend/graphql'
 import { ErrorBoundary } from './components/ErrorBoundary'
 import { FeedbackText } from './components/FeedbackText'
 import { HeroPage } from './components/HeroPage'
@@ -41,8 +38,8 @@ import { LayoutRouteProps } from './routes'
 import { SiteAdminAreaRoute } from './site-admin/SiteAdminArea'
 import { SiteAdminSideBarGroups } from './site-admin/SiteAdminSidebar'
 import { ThemePreference } from './theme'
+import { ActivationStatus } from './tracking/ActivationStatus'
 import { eventLogger } from './tracking/eventLogger'
-import { logUserEvent } from './user/account/backend'
 import { UserAccountAreaRoute } from './user/account/UserAccountArea'
 import { UserAccountSidebarItems } from './user/account/UserAccountSidebar'
 import { UserAreaRoute } from './user/area/UserArea'
@@ -66,7 +63,11 @@ export interface SourcegraphWebAppProps extends KeybindingsProps {
     routes: ReadonlyArray<LayoutRouteProps>
 }
 
-interface SourcegraphWebAppState extends PlatformContextProps, SettingsCascadeProps, ExtensionsControllerProps {
+interface SourcegraphWebAppState
+    extends PlatformContextProps,
+        SettingsCascadeProps,
+        ExtensionsControllerProps,
+        ActivationProps {
     error?: Error
 
     /** The currently authenticated user (or null if the viewer is anonymous). */
@@ -87,17 +88,6 @@ interface SourcegraphWebAppState extends PlatformContextProps, SettingsCascadePr
      * The current search query in the navbar.
      */
     navbarSearchQuery: string
-
-    /**
-     * Defines the activation steps that a user must complete and where
-     * this is fetched from.
-     */
-    activation?: ActivationParams
-
-    /**
-     * Specifies which activation steps have been completed.
-     */
-    activationCompleted?: ActivationCompleted
 }
 
 const LIGHT_THEME_LOCAL_STORAGE_KEY = 'light-theme'
@@ -162,12 +152,41 @@ export class SourcegraphWebApp extends React.Component<SourcegraphWebAppProps, S
                 authenticatedUser => {
                     this.setState({
                         authenticatedUser,
-                        activation: this.getActivationState(authenticatedUser),
                     })
                 },
-                () => this.setState({ authenticatedUser: null, activation: undefined, activationCompleted: undefined })
+                () =>
+                    this.setState({
+                        authenticatedUser: null,
+                    })
             )
         )
+
+        if (!window.context.sourcegraphDotComMode) {
+            // Add activation status
+            this.subscriptions.add(
+                authenticatedUser
+                    .pipe(
+                        map(user => (user ? new ActivationStatus(user) : undefined)),
+                        switchMap(activationStatus => {
+                            if (!activationStatus) {
+                                return of(undefined)
+                            }
+                            return activationStatus.completed.pipe(
+                                map(completed => {
+                                    const activation: Activation = {
+                                        steps: activationStatus.steps,
+                                        completed: completed || undefined,
+                                        update: activationStatus.updateCompleted,
+                                        refetch: activationStatus.refetchCompleted,
+                                    }
+                                    return activation
+                                })
+                            )
+                        })
+                    )
+                    .subscribe(activation => this.setState({ activation }))
+            )
+        }
 
         this.subscriptions.add(
             combineLatest(from(this.state.platformContext.settings), authenticatedUser.pipe(startWith(null))).subscribe(
@@ -206,62 +225,6 @@ export class SourcegraphWebApp extends React.Component<SourcegraphWebAppProps, S
                 this.setState({ systemIsLightTheme: !event.matches })
             })
         )
-    }
-
-    private getActivationState = (authenticatedUser: GQL.IUser | null): ActivationParams | undefined => {
-        if (window.context.sourcegraphDotComMode || !authenticatedUser) {
-            return
-        }
-        const fetcher = fetchActivationStatus(authenticatedUser.siteAdmin)
-        const firstFetched = new Promise<void>(resolve => {
-            fetcher()
-                .pipe(first())
-                .subscribe(completed => {
-                    this.setState({ activationCompleted: completed })
-                    resolve()
-                })
-        })
-        return {
-            steps: getActivationSteps(authenticatedUser.siteAdmin),
-            fetcher,
-            firstFetched,
-        }
-    }
-
-    private refetchActivation = () => {
-        if (!this.state.activation) {
-            return
-        }
-        this.state.activation
-            .fetcher()
-            .pipe(first())
-            .subscribe(c => this.setState({ activationCompleted: c }))
-    }
-
-    private updateActivation = (update: ActivationCompleted) => {
-        if (!this.state.activation) {
-            return
-        }
-        this.state.activation.firstFetched.then(() => {
-            if (!this.state.activation) {
-                return
-            }
-
-            // Send update to server for events that don't themselves trigger
-            // an update.
-            if (update.FoundReferences) {
-                logUserEvent(GQL.UserEvent.CODEINTELREFS)
-            }
-
-            const newVal: ActivationCompleted = {}
-            Object.assign(newVal, this.state.activationCompleted)
-            for (const step of this.state.activation.steps) {
-                if (update[step.id] !== undefined) {
-                    newVal[step.id] = update[step.id]
-                }
-            }
-            this.setState({ activationCompleted: newVal })
-        })
     }
 
     public componentWillUnmount(): void {
@@ -335,15 +298,7 @@ export class SourcegraphWebApp extends React.Component<SourcegraphWebAppProps, S
                                         platformContext={this.state.platformContext}
                                         extensionsController={this.state.extensionsController}
                                         // Activation
-                                        activation={
-                                            this.state.activation &&
-                                            this.state.activationCompleted && {
-                                                steps: this.state.activation.steps,
-                                                update: this.updateActivation,
-                                                refetch: this.refetchActivation,
-                                                completed: this.state.activationCompleted,
-                                            }
-                                        }
+                                        activation={this.state.activation}
                                     />
                                 )}
                             />
@@ -363,144 +318,4 @@ export class SourcegraphWebApp extends React.Component<SourcegraphWebAppProps, S
     private onNavbarQueryChange = (navbarSearchQuery: string) => {
         this.setState({ navbarSearchQuery })
     }
-}
-
-interface ActivationParams {
-    fetcher: () => Observable<ActivationCompleted>
-    firstFetched: Promise<void>
-    steps: ActivationStep[]
-}
-
-const fetchActivationStatus = (isSiteAdmin: boolean) => () =>
-    queryGraphQL(
-        isSiteAdmin
-            ? gql`
-                  query {
-                      externalServices {
-                          totalCount
-                      }
-                      repositories(enabled: true) {
-                          totalCount
-                      }
-                      viewerSettings {
-                          final
-                      }
-                      users {
-                          totalCount
-                      }
-                      currentUser {
-                          usageStatistics {
-                              searchQueries
-                              findReferencesActions
-                          }
-                      }
-                  }
-              `
-            : gql`
-                  query {
-                      currentUser {
-                          usageStatistics {
-                              searchQueries
-                              findReferencesActions
-                          }
-                      }
-                  }
-              `
-    ).pipe(
-        map(dataOrThrowErrors),
-        map(data => {
-            const authProviders = window.context.authProviders
-            const completed: ActivationCompleted = {
-                DidSearch: !!data.currentUser && data.currentUser.usageStatistics.searchQueries > 0,
-                FoundReferences: !!data.currentUser && data.currentUser.usageStatistics.findReferencesActions > 0,
-            }
-            if (isSiteAdmin) {
-                completed.ConnectedCodeHost = data.externalServices && data.externalServices.totalCount > 0
-                completed.EnabledRepository =
-                    data.repositories && data.repositories.totalCount !== null && data.repositories.totalCount > 0
-                if (authProviders) {
-                    completed.EnabledSharing =
-                        data.users.totalCount > 1 || authProviders.filter(p => !p.isBuiltin).length > 0
-                }
-            }
-            return completed
-        })
-    )
-
-const fetchReferencesLink = (): Observable<string | null> =>
-    queryGraphQL(gql`
-        query {
-            repositories(enabled: true, cloned: true, first: 100, indexed: true) {
-                nodes {
-                    url
-                    gitRefs {
-                        totalCount
-                    }
-                }
-            }
-        }
-    `).pipe(
-        map(dataOrThrowErrors),
-        map(data => {
-            if (!data.repositories.nodes) {
-                return null
-            }
-            const repositoryURLs = data.repositories.nodes
-                .filter(r => r.gitRefs && r.gitRefs.totalCount > 0)
-                .sort((r1, r2) => r2.gitRefs!.totalCount! - r1.gitRefs!.totalCount)
-                .map(r => r.url)
-            if (repositoryURLs.length === 0) {
-                return null
-            }
-            return repositoryURLs[0]
-        })
-    )
-
-const getActivationSteps = (isSiteAdmin: boolean): ActivationStep[] => {
-    const sources: (ActivationStep & { siteAdminOnly?: boolean })[] = [
-        {
-            id: 'ConnectedCodeHost',
-            title: 'Connect your code host',
-            detail: 'Configure Sourcegraph to talk to your code host and fetch a list of your repositories.',
-            link: { to: '/site-admin/external-services' },
-            siteAdminOnly: true,
-        },
-        {
-            id: 'EnabledRepository',
-            title: 'Enable repositories',
-            detail: 'Select which repositories Sourcegraph should pull and index from your code host(s).',
-            link: { to: '/site-admin/repositories' },
-            siteAdminOnly: true,
-        },
-        {
-            id: 'DidSearch',
-            title: 'Search your code',
-            detail: 'Perform a search query on your code.',
-            link: { to: '/search' },
-        },
-        {
-            id: 'FoundReferences',
-            title: 'Find some references',
-            detail:
-                'To find references of a token, navigate to a code file in one of your repositories, hover over a token to activate the tooltip, and then click "Find references".',
-            onClick: (event: React.MouseEvent<HTMLElement>, history: H.History) =>
-                fetchReferencesLink()
-                    .pipe(first())
-                    .subscribe(r => {
-                        if (r) {
-                            history.push(r)
-                        } else {
-                            alert('Must add repositories before finding references')
-                        }
-                    }),
-        },
-        {
-            id: 'EnabledSharing',
-            title: 'Configure SSO or share with teammates',
-            detail: 'Configure a single-sign on (SSO) provider or have at least one other teammate sign up.',
-            link: { to: 'https://docs.sourcegraph.com/admin/auth', target: '_blank' },
-            siteAdminOnly: true,
-        },
-    ]
-    return sources.filter(e => true || !e.siteAdminOnly).map(({ siteAdminOnly, ...step }) => step)
 }
